Refactored CatalogNavigationItem as Domen Element Transfer files
