--- src/components/CatalogNavigationItem/index.js
@@ -1,3 +1,0 @@
-import CatalogNavigationItem from './CatalogNavigationItem'
-
-export default CatalogNavigationItem
